feat: refactor filters
